moved all platform integrations to integration module
